+++ frappe/components/Icons/MaximizeIcon.vue
