+++ frappe/components/Icons/MinimizeIcon.vue
