+++ frappe/components/Icons/StepsIcon.vue
